+++ extract_msg/__init__.py
@@ -27,8 +27,8 @@ https://github.com/mattgwwalker/msg-extractor
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 __author__ = 'The Elemental of Destruction & Matthew Walker'
-__date__ = '2020-08-13'
-__version__ = '0.26.4'
+__date__ = '2020-08-22'
+__version__ = '0.27.0'
 
 import logging