+++ extract_msg/prop.py
@@ -1,7 +1,8 @@
+import datetime
 import logging
 
 from extract_msg import constants
-from extract_msg.utils import properHex
+from extract_msg.utils import fromTimeStamp, msgEpoch, properHex
 
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())
@@ -123,8 +124,7 @@ class FixedLengthProp(PropBase):
             value = (constants.STI64.unpack(value))[0] / 10000.0
         elif _type == 0x0007:  # PtypFloatingTime
             value = constants.STF64.unpack(value)[0]
-            # TODO parsing for this
-            pass
+            return constants.PYTPFLOATINGTIME_START + datetime.timedelta(days = value)
         elif _type == 0x000A:  # PtypErrorCode
             value = constants.STI32.unpack(value)[0]
             # TODO parsing for this
@@ -134,7 +134,13 @@ class FixedLengthProp(PropBase):
         elif _type == 0x0014:  # PtypInteger64
             value = constants.STI64.unpack(value)[0]
         elif _type == 0x0040:  # PtypTime
-            value = constants.ST3.unpack(value)[0]
+            try:
+                value = fromTimeStamp(msgEpoch(constants.ST3.unpack(value)[0]))
+            except Exception as e:
+                logger.exception(e)
+                logger.error('Timestamp value of {} caused an exception. This was probably caused by the time stamp being too far in the future.')
+                print(self.raw)
+                value = constants.ST3.unpack(value)[0]
         elif _type == 0x0048:  # PtypGuid
             # TODO parsing for this
             pass