+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+**v0.27.0**
+* [[mattgwwalker #143](https://github.com/mattgwwalker/msg-extractor/issues/143)] Added new class `Appointment` that can handle outlook appointments or meetings.
+* [[mattgwwalker #143](https://github.com/mattgwwalker/msg-extractor/issues/143)] Added new class `MessageBase` for classes that end up being mostly like the `Message` class. Currently subclassed by `Message` and `Appointment`. This should not have a direct effect on any code that uses this module.
+* Added support for `PtypFloatingTime` in `utils.parseType`.
+* Added proper support for `PtypTime` in `FixedLengthProp.parseType`
+
 **v0.26.4**
 * Added new function `MSGFile._ensureSetNamed` which acts like `MSGFile._ensureSet` except that it works with named properties.
 * Added a version of the previous function to the `Attachment` and `Recipient` classes.